Move FDB status checks into dedicated package for easier usability
