+++ pkg/fdbstatus/fdb_status_helper.go
@@ -18,7 +18,7 @@
  * limitations under the License.
  */
 
-package internal
+package fdbstatus
 
 import (
 	"fmt"
