+++ pkg/fdbstatus/fdb_status_helper_test.go
@@ -18,7 +18,7 @@
  * limitations under the License.
  */
 
-package internal
+package fdbstatus
 
 import (
 	"fmt"
